Move circle progressbars to its own component file
