+++ src/components/EditData.js
@@ -14,13 +14,10 @@ import {
 
 import {
     LandingPageContainer,
-    ProgressBarsContainer,
-    ProgressbarSleepAmount,
     ButtonsContainer,
     InputFormButton,
-    Emoji,
 } from "./UserLandingPage/LandingPage";
-import CircleProgressbar from "./UserLandingPage/CircleProgressbar";
+import CircleProgressbars from "./UserLandingPage/CircleProgressbars";
 import UserInputForm from "./UserLandingPage/UserInputForm";
 
 const DateInputContainer = styled.div`
@@ -39,10 +36,6 @@ const DateInputContainer = styled.div`
     }
 `;
 
-const ProgressCirclesContainer = styled(ProgressBarsContainer)`
-    margin: 10px 0;
-`;
-
 const EditData = ({
     dateToEdit,
     getDataFromOneDate, // need to destructure b/c of useEffect dependency array warning error
@@ -52,10 +45,6 @@ const EditData = ({
     ...props
 }) => {
     const [date, setDate] = useState(new Date());
-    const [averageMoodEmoji, setAverageMoodEmoji] = useState(0);
-    const [moodEmojiAriaLabel, setMoodEmojiAriaLabel] = useState("");
-    const [averageTirednessEmoji, setAverageTirednessEmoji] = useState(0);
-    const [tirednessEmojiAriaLabel, setTirednessEmojiAriaLabel] = useState("");
 
     // for "closing" forms when clicking outside of them
     const wakeUpFormRef = useRef(null);
@@ -70,16 +59,6 @@ const EditData = ({
     const [wakeUpSlide, setWakeUpSlide] = useState(0);
     const [bedtimeSlide, setBedtimeSlide] = useState(0);
     const [middaySlide, setMiddaySlide] = useState(0);
-
-    // initial values for the times, moods, and tiredness before doing a GET request
-    // const [initialValues, setInitialValues] = useState({
-    //     mood: 0,
-    //     tiredness: 0,
-    // });
-    // const [initialValuesPlusTime, setInitialValuesPlusTime] = useState({
-    //     ...initialValues,
-    //     time: "",
-    // });
 
     // handleTap functions for the IconTab animations
     const wakeUpTap = () => {
@@ -130,26 +109,12 @@ const EditData = ({
         setBedtimeSlide(0);
     });
 
-    const setProgressBarColor = percentage => {
-        if (percentage < 34) {
-            return "#F20000"; //red
-        } else if (percentage < 66 && percentage > 34) {
-            return "#EFD914"; // yellow
-        } else if (percentage > 66) {
-            return "#20C261"; // green
-        }
-
-        return "#20C261"; // green
-    };
-
     const handleDateChange = e => {
         setDate(e.target.value);
     };
 
     // handleSubmit for sending the PUT request to the server
     const handleMoodSubmit = (timeOfDay, dateId, updatedMood) => {
-        // e.preventDefault();
-
         // invoke action creator here that does the PUT request
         // action.payload = the date's updated redux state for the times, mood, and tiredness
         props.editMood(timeOfDay, dateId, updatedMood);
@@ -161,14 +126,10 @@ const EditData = ({
         dateId,
         updatedTiredness
     ) => {
-        // e.preventDefault();
-        console.log("initialValues in tiredness submit: ", initialVals);
         props.editTiredness(timeOfDay, dateId, updatedTiredness);
     };
 
     const handleSleepTimesSubmit = (timeOfDay, dateId, updatedTime) => {
-        // e.preventDefault();
-
         // updatedTime needs to be Date.toISOString()
         props.editWakeAndBedTimes(timeOfDay, dateId, updatedTime);
     };
@@ -180,53 +141,10 @@ const EditData = ({
             getDataFromOneDate(day);
         };
 
-        console.log("date: ", formatDate(date));
+        // console.log("date: ", formatDate(date));
 
         getDateData(formatDate(date));
     }, [date, getDataFromOneDate]);
-
-    useEffect(() => {
-        let moodEmoji = "";
-        let moodAriaLabel = "";
-        let tirednessEmoji = "";
-        let tirednessAriaLabel = "";
-        const { bad: badMood, ok: okMood, great: greatMood } = moodEmojis;
-        const {
-            bad: badTiredness,
-            ok: okTiredness,
-            great: greatTiredness,
-        } = tirednessEmojis;
-
-        if (dateToEdit.averageMood <= 1) {
-            moodEmoji = badMood.emoji;
-            moodAriaLabel = badMood.desc;
-        } else if (dateToEdit.averageMood > 1 && dateToEdit.averageMood <= 2) {
-            moodEmoji = okMood.emoji;
-            moodAriaLabel = okMood.desc;
-        } else if (dateToEdit.averageMood > 2) {
-            moodEmoji = greatMood.emoji;
-            moodAriaLabel = greatMood.desc;
-        }
-
-        if (dateToEdit.averageTiredness <= 1) {
-            tirednessEmoji = greatTiredness.emoji;
-            tirednessAriaLabel = greatTiredness.desc;
-        } else if (
-            dateToEdit.averageTiredness > 1 &&
-            dateToEdit.averageTiredness <= 2
-        ) {
-            tirednessEmoji = okTiredness.emoji;
-            tirednessAriaLabel = okTiredness.desc;
-        } else if (dateToEdit.averageTiredness > 2) {
-            tirednessEmoji = badTiredness.emoji;
-            tirednessAriaLabel = badTiredness.desc;
-        }
-
-        setAverageMoodEmoji(moodEmoji);
-        setMoodEmojiAriaLabel(moodAriaLabel);
-        setAverageTirednessEmoji(tirednessEmoji);
-        setTirednessEmojiAriaLabel(tirednessAriaLabel);
-    }, [dateToEdit, moodEmojis, tirednessEmojis]);
 
     useEffect(() => {
         // when the input values get updated in the redux store (pass into the dependency array for this useEffect), set the local state 'initialValues' and 'initialValuesPlusTime'
@@ -254,40 +172,7 @@ const EditData = ({
             </DateInputContainer>
 
             {/* Amount of sleep, mood, and tiredness levels compared to averages as circular progress "bars" */}
-            <ProgressCirclesContainer>
-                {/* sleep ratio */}
-                <CircleProgressbar
-                    progressColor={setProgressBarColor(30)}
-                    value={dateToEdit.totalTimeInBed} // need to divide by overall average
-                >
-                    {/* placeholder value */}
-                    <ProgressbarSleepAmount>
-                        {dateToEdit.totalTimeInBed}hrs
-                    </ProgressbarSleepAmount>
-                </CircleProgressbar>
-
-                {/* mood ratio */}
-                <CircleProgressbar
-                    progressColor={setProgressBarColor(52)}
-                    value={dateToEdit.averageMood} // need to divide by overall average
-                >
-                    <Emoji
-                        emoji={averageMoodEmoji}
-                        ariaLabel={moodEmojiAriaLabel}
-                    />
-                </CircleProgressbar>
-
-                {/* tiredness ratio */}
-                <CircleProgressbar
-                    progressColor={setProgressBarColor(80)}
-                    value={dateToEdit.averageTiredness} // need to divide by overall average
-                >
-                    <Emoji
-                        emoji={averageTirednessEmoji}
-                        ariaLabel={tirednessEmojiAriaLabel}
-                    />
-                </CircleProgressbar>
-            </ProgressCirclesContainer>
+            <CircleProgressbars />
 
             <ButtonsContainer>
                 <InputFormButton
